ai_study/backend/python_agent/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
openai>=1.0.0
pytest>=8.0.0
